config: move CoS warn validators to own file
Pure code-motion for #4845. Move the class-of-service WARN validators
(validateCoSOversubscriptionWarnings, classOfServiceClassifierQueueWarnings)
verbatim from compiler_validate_warn.go into compiler_validate_warn_cos.go.
This drops the remainder file below the ~2,000-LOC modularity threshold.

Bodies are byte-identical; only the file each function lives in and the
per-file import block change. Same package config, same signatures. No
behavior change.
